Move legacy tests into tests directory
